tests: utils: share fillmem guest memory dirtying helper
test_snapshot_restore_cross_kernel.py imported make_guest_dirty_memory
from the balloon test module, and test_hotplug_memory.py carries its own
fillmem invocation.

Move the make_guest_dirty_memory helper next to get_stable_rss_mem in
framework.utils so that every caller can share one implementation. No
functional change.
